requirements.txt
Flask
python-dotenv
requests
signalwire-swaig
flask-httpauth